format

+++ tests/Neo.UnitTests/IO/Caching/UT_HashSetUInt256Cache.cs
@@ -1,6 +1,6 @@
 // Copyright (C) 2015-2025 The Neo Project.
 //
-// UT_HashSetCache.cs file belongs to the neo project and is free
+// UT_HashSetUInt256Cache.cs file belongs to the neo project and is free
 // software distributed under the MIT software license, see the
 // accompanying file LICENSE in the main directory of the
 // repository or http://www.opensource.org/licenses/mit-license.php
@@ -25,7 +25,7 @@ namespace Neo.UnitTests.IO.Caching
         private static UInt256 ToHash(byte num)
         {
             var hash = new byte[32];
-            hash[0] = (byte)num;
+            hash[0] = num;
             return new UInt256(hash);
         }
